Refactor: Move checkBurningManSelectionHeight to DelayedPayoutTxValidation
